Potential fix for code scanning alert no. 40: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger type before the result is assigned to the size_t variable. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using the larger type, preventing overflow.

Specifically, we will cast width to size_t before multiplying it by height on line 118. This change will ensure that the multiplication is performed using size_t, which can hold larger values and prevent overflow.

+++ libwebp-1.0.3/src/enc/alpha_enc.c
@@ -115,7 +115,7 @@ static int EncodeAlphaInternal(const uint8_t* const data, int width, int height,
   const uint8_t* alpha_src;
   WebPFilterFunc filter_func;
   uint8_t header;
-  const size_t data_size = width * height;
+  const size_t data_size = (size_t)width * height;
   const uint8_t* output = NULL;
   size_t output_size = 0;
   VP8LBitWriter tmp_bw;
